Finished Vileplume Alg sorta..?
I completed the Vileplume program, but it's a very bad version of it.. I
still need to refine the algorithm a lot, but this is what I have for
the time being.

There is also a glitch in the Basics Checker program that doesn't group
alike Pokemon, which I am well aware of.